+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Update `cmd/hidden_lake/messenger`: delete QR codes
 - Update `cmd/hidden_lake/messenger`: append own address in settings
+- Update `cmd/hidden_lake/messenger`: escapeOutput from JS -> html.EscapeString from Go
 
 ### BUG FIXES